requirements.txt
# VLA Project Requirements
# ========================

# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0

# Transformers and VLM
transformers>=4.37.0
accelerate>=0.25.0
peft>=0.7.0  # For LoRA
bitsandbytes>=0.41.0  # For quantization (optional)

# Qwen VL dependencies
qwen-vl-utils>=0.0.2
tiktoken>=0.5.0

# Data processing
numpy>=1.24.0
pillow>=10.0.0
h5py>=3.9.0
opencv-python>=4.8.0

# Training utilities
tqdm>=4.65.0
wandb>=0.15.0  # Optional: for experiment tracking
tensorboard>=2.14.0  # Optional: alternative logging

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Isaac Lab (installed separately)
# See: https://isaac-sim.github.io/IsaacLab/main/source/setup/installation.html

# Development
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

# Optional: for faster data loading
# lmdb>=1.4.0
# webdataset>=0.2.0
